spsearch/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
sentence-transformers
pgvector
